analytics/requirements.txt
prefect>=2.19.0
google-cloud-bigquery>=3.20.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
httpx>=0.27.0
ruff>=0.4.0
